+++ CHANGELOG.md
@@ -14,19 +14,28 @@ You may also find the [Upgrade Guide](https://rust-random.github.io/book/update.
 - Added a `serde1` feature and added Serialize/Deserialize to `UniformInt` and `WeightedIndex` (#974)
 - Document types supported by `random` (#994)
 - Implement weighted sampling without replacement (#976, #1013)
+- Add `IteratorRandom::choose_stable` as an alternative to `choose` which does not depend on size hints (#1057)
 
 ### Changes
+- `getrandom` updated to v0.2 (#1041)
+- `ThreadRng` is no longer `Copy` to enable safe usage within thread-local destructors (see #968)
 - `gen_range(a, b)` was replaced with `gen_range(a..b)`, and `gen_range(a..=b)`
   is supported (#744, #1003). Note that `a` and `b` can no longer be references or SIMD types.
 - Replace `AsByteSliceMut` with `Fill` (#940)
 - Move alias method for `WeightedIndex` to `rand_distr` (#945)
 - `Alphanumeric` samples bytes instead of chars (#935)
 - The minimum supported Rust version is now 1.36 (#1011)
+- Restrict `rand::rngs::adapter` to `std` (#1027)
 - Better NaN handling for `WeightedIndex` (#1005)
 - Implement `IntoIterator` for `IndexVec`, replacing the `into_iter` method (#1007)
 - Reduce packaged crate size (#983)
-- Drop some unsafe code (#962, #963)
+- Drop some unsafe code (#962, #963, #1011)
 - Improve treatment of rounding errors in `WeightedIndex::update_weights` (#956)
+- `StdRng`: Switch from ChaCha20 to ChaCha12 for better performance (#1028)
+- `SmallRng`: Replace PCG algorithm with xoshiro{128,256}++ (#1038)
+- The `nightly` feature no longer implies the `simd_support` feature (#1048)
+- Fix `simd_support` feature to work on current nightlies (#1056)
+- Improve accuracy and performance of `IteratorRandom::choose` (#1059)
 
 ## [0.7.3] - 2020-01-10
 ### Fixes